+++ Sources/MMIO/MMIOMacros.swift
@@ -10,42 +10,127 @@
 //===----------------------------------------------------------------------===//
 
 // Deprecated RegisterBank macros
-@available(*, deprecated, renamed: "RegisterBlock()")
+@_documentation(visibility: internal)
+@available(
+  *, deprecated, message: "Use @RegisterBlock() instead.",
+  renamed: "RegisterBlock()"
+)
 @attached(member, names: named(unsafeAddress), named(init), named(interposer))
 public macro RegisterBank() =
   #externalMacro(module: "MMIOMacros", type: "RegisterBlockMacro")
 
-@available(*, deprecated, renamed: "RegisterBlock(offset:)")
+@_documentation(visibility: internal)
+@available(
+  *, deprecated, message: "Use @RegisterBlock(offset:) instead.",
+  renamed: "RegisterBlock(offset:)"
+)
 @attached(accessor)
 public macro RegisterBank(offset: Int) =
   #externalMacro(module: "MMIOMacros", type: "RegisterBlockScalarMemberMacro")
 
-@available(*, deprecated, renamed: "RegisterBlock(offset:stride:count:)")
+@_documentation(visibility: internal)
+@available(
+  *, deprecated, message: "Use @RegisterBlock(offset:stride:count:) instead.",
+  renamed: "RegisterBlock(offset:stride:count:)"
+)
 @attached(accessor)
 public macro RegisterBank(offset: Int, stride: Int, count: Int) =
   #externalMacro(module: "MMIOMacros", type: "RegisterBlockArrayMemberMacro")
 
-// RegisterBlock macros
+// MARK: - RegisterBlock macros
+
+/// Defines a group of memory-mapped registers, such as a hardware peripheral or
+/// a logical section within it.
+///
+/// Attach this macro to a `struct` to generate essential properties for MMIO
+/// operations, including `unsafeAddress` for the base address and an
+/// initializer. Members within this struct that represent individual registers
+/// or nested register blocks must be annotated with ``RegisterBlock(offset:)``
+/// for single elements or ``RegisterBlock(offset:stride:count:)`` for arrays of
+/// elements.
+///
+/// For a comprehensive guide on defining peripheral layouts, see
+/// <doc:Register-Layouts>.
 @attached(member, names: named(unsafeAddress), named(init), named(interposer))
 @attached(extension, conformances: RegisterProtocol)
 public macro RegisterBlock() =
   #externalMacro(module: "MMIOMacros", type: "RegisterBlockMacro")
 
+/// Defines a single register or a nested register block within a parent
+/// ``MMIO/RegisterBlock()``.
+///
+/// Use this macro to structure individual hardware registers or cohesive
+/// sub-units within a larger peripheral definition.
+///
+/// - Parameter offset: The byte offset of this register or register block,
+///   calculated from the base address of the enclosing
+///   ``MMIO/RegisterBlock()``.
+///
+/// For defining arrays of registers or register blocks, use
+/// ``MMIO/RegisterBlock(offset:stride:count:)``.
 @attached(accessor)
 public macro RegisterBlock(offset: Int) =
   #externalMacro(module: "MMIOMacros", type: "RegisterBlockScalarMemberMacro")
 
+/// Defines an array of registers or nested register blocks within a parent
+/// ``MMIO/RegisterBlock()``.
+///
+/// This macro is used for defining repetitive hardware structures, such as
+/// multiple identical channels or a series of configuration registers. The
+/// annotated property should be of type ``MMIO/RegisterArray``.
+///
+/// - Parameters:
+///   - offset: The byte offset of the first element in the array,
+///     relative to the base address of the enclosing ``MMIO/RegisterBlock()``.
+///   - stride: The byte distance from the start of one element in the
+///     array to the start of the next. This is often the size of the element
+///     type but can be larger if there's padding between elements in the
+///     hardware memory layout.
+///   - count: The total number of elements in the array.
 @attached(accessor)
 public macro RegisterBlock(offset: Int, stride: Int, count: Int) =
   #externalMacro(module: "MMIOMacros", type: "RegisterBlockArrayMemberMacro")
 
-// Register macros
+// MARK: - Register macros
+/// Defines the bit-level layout of an individual hardware register.
+///
+/// Attach this macro to a `struct`. The struct contains properties
+/// that define individual bit fields, using macros such as
+/// ``MMIO/ReadWrite(bits:as:)``, ``MMIO/ReadOnly(bits:as:)``,
+/// ``MMIO/WriteOnly(bits:as:)``, and ``MMIO/Reserved(bits:as:)``. This macro
+/// automatically makes the struct conform to the ``MMIO/RegisterValue``
+/// protocol, enabling its use with ``MMIO/Register`` instances.
+///
+/// - Parameter bitWidth: The total width of the register in bits (e.g., 8, 16,
+///   32, or 64). This value must accurately reflect the size of the actual
+///   hardware register.
+///
+/// For detailed usage examples and explanations, refer to
+/// <doc:Register-Layouts> for structuring registers and
+/// <doc:Bit-Fields> for defining their contents.
 @attached(member, names: arbitrary)
 @attached(memberAttribute)
 @attached(extension, conformances: RegisterValue)
 public macro Register(bitWidth: Int) =
   #externalMacro(module: "MMIOMacros", type: "RegisterMacro")
 
+/// Defines reserved bit(s) within a hardware register.
+///
+/// Reserved bits are parts of a register that are typically unused, have
+/// hardware-defined behavior, or are set aside for future use by the hardware
+/// vendor. Software should generally avoid writing to reserved bits unless
+/// explicitly instructed by the hardware documentation.
+///
+/// - Parameters:
+///   - bits: A `RangeExpression` (e.g., `0..<4`) or a comma-separated list
+///     of range expressions (for <doc:Discontiguous-Bit-Fields>)
+///     specifying the bit range(s) occupied by these reserved bits. Use `...`
+///     to indicate that the reserved field spans the entire register width.
+///   - as: An optional type conforming to ``BitFieldProjectable`` for
+///     <doc:Type-Projections>. This is typically not used for
+///     reserved fields, defaulting to `Never`.
+///
+/// - SeeAlso: <doc:Bit-Fields>.
 @attached(accessor)
 public macro Reserved<Range, Value>(
   bits: Range..., as: Value.Type = Never.self
@@ -54,6 +139,7 @@ public macro Reserved<Range, Value>(
 where
   Range: RangeExpression, Range.Bound: BinaryInteger, Value: BitFieldProjectable
 
+@_documentation(visibility: internal)
 @attached(accessor)
 public macro Reserved<Value>(
   bits: UnboundedRange, as: Value.Type = Never.self
@@ -61,6 +147,21 @@ public macro Reserved<Value>(
   #externalMacro(module: "MMIOMacros", type: "ReservedMacro")
 where Value: BitFieldProjectable
 
+/// Defines a read-write bit field within a register.
+///
+/// Bits within this field can be both read from and written to by software.
+///
+/// - Parameters:
+///   - bits: A `RangeExpression` (e.g., `0..<4` for bits 0, 1, 2, and 3)
+///     or a comma-separated list of such expressions to define a
+///     <doc:Discontiguous-Bit-Fields>. Use `...` if the field spans the entire
+///     register.
+///   - as: An optional type conforming to ``BitFieldProjectable`` to
+///     enable <doc:Type-Projections>. For example, can use `Bool`.self for
+///     single-bit flags, or a custom enumeration for multi-bit fields
+///     representing specific states.
+///
+/// For more information, see <doc:Bit-Fields>.
 @attached(accessor)
 public macro ReadWrite<Range, Value>(
   bits: Range..., as: Value.Type = Never.self
@@ -69,6 +170,7 @@ public macro ReadWrite<Range, Value>(
 where
   Range: RangeExpression, Range.Bound: BinaryInteger, Value: BitFieldProjectable
 
+@_documentation(visibility: internal)
 @attached(accessor)
 public macro ReadWrite<Value>(
   bits: UnboundedRange, as: Value.Type = Never.self
@@ -76,6 +178,21 @@ public macro ReadWrite<Value>(
   #externalMacro(module: "MMIOMacros", type: "ReadWriteMacro")
 where Value: BitFieldProjectable
 
+/// Defines a read-only bit field within a register.
+///
+/// The value of this field is determined by the hardware and can only be read
+/// by software. Attempting to write to a read-only field typically has no
+/// effect or is an error, depending on the hardware.
+///
+/// - Parameters:
+///   - bits: A `RangeExpression` (e.g., `0..<4`) or a comma-separated list
+///     of such expressions (for <doc:Discontiguous-Bit-Fields>) specifying the
+///     bit range(s). Use `...` if the field spans the entire
+///     register.
+///   - as: An optional type conforming to ``BitFieldProjectable`` for
+///     <doc:Type-Projections>.
+///
+/// - SeeAlso: <doc:Bit-Fields>.
 @attached(accessor)
 public macro ReadOnly<Range, Value>(
   bits: Range..., as: Value.Type = Never.self
@@ -84,6 +201,7 @@ public macro ReadOnly<Range, Value>(
 where
   Range: RangeExpression, Range.Bound: BinaryInteger, Value: BitFieldProjectable
 
+@_documentation(visibility: internal)
 @attached(accessor)
 public macro ReadOnly<Value>(
   bits: UnboundedRange, as: Value.Type = Never.self
@@ -91,6 +209,20 @@ public macro ReadOnly<Value>(
   #externalMacro(module: "MMIOMacros", type: "ReadOnlyMacro")
 where Value: BitFieldProjectable
 
+/// Defines a write-only bit field within a register.
+///
+/// This field can only be written to by software. Reading from a write-only
+/// field might return an undefined value, always read as zero, or trigger
+/// specific hardware side effects, as defined by the hardware documentation.
+///
+/// - Parameters:
+///   - bits: A `RangeExpression` (e.g., `0..<4`) or a comma-separated list
+///     of such expressions (for <doc:Discontiguous-Bit-Fields>) specifying the
+///     bit range(s). Use `...` if the field spans the entire register.
+///   - as: An optional type conforming to ``BitFieldProjectable`` for
+///     <doc:Type-Projections>.
+///
+/// - SeeAlso: <doc:Bit-Fields>.
 @attached(accessor)
 public macro WriteOnly<Range, Value>(
   bits: Range..., as: Value.Type = Never.self
@@ -99,6 +231,7 @@ public macro WriteOnly<Range, Value>(
 where
   Range: RangeExpression, Range.Bound: BinaryInteger, Value: BitFieldProjectable
 
+@_documentation(visibility: internal)
 @attached(accessor)
 public macro WriteOnly<Value>(
   bits: UnboundedRange, as: Value.Type = Never.self
