+++ Sources/MMIO/RegisterProtocol.swift
@@ -9,12 +9,38 @@
 //
 //===----------------------------------------------------------------------===//
 
-// This only exists because RegisterBlock is not a concrete type, ideally
-// RegisterBlock and Register would both be the same Register type
+/// A protocol defining the common initialization interface for memory-mapped I/O entities.
+///
+/// This protocol is primarily adopted by `struct`s that are annotated with the
+/// ``RegisterBlock()`` macro. It ensures that types representing a block of
+/// registers (a peripheral or a sub-block) can be initialized with a base
+/// memory address and, optionally, an ``MMIOInterposer`` for testing purposes.
+///
+/// The main use of `RegisterProtocol` is to enable generic handling of register
+/// blocks, particularly within ``RegisterArray`` when it stores an array of
+/// register blocks rather than individual registers.
+///
+/// > Note: You typically do not interact with this protocol directly or
+///   conform types to it manually. Conformance is automatically provided by the
+///   ``RegisterBlock()`` macro.
 public protocol RegisterProtocol {
   #if FEATURE_INTERPOSABLE
+  /// Initializes a new instance of the MMIO entity (e.g., a register block).
+  ///
+  /// This initializer is used when the `FEATURE_INTERPOSABLE` flag is enabled
+  /// during compilation, allowing an ``MMIOInterposer`` to be provided for testing.
+  ///
+  /// - Parameters:
+  ///   - unsafeAddress: The base memory address for this entity.
+  ///   - interposer: An optional ``MMIOInterposer`` to intercept
+  ///     memory accesses. If `nil`, accesses go directly to hardware.
   init(unsafeAddress: UInt, interposer: (any MMIOInterposer)?)
   #else
+  /// Initializes a new instance of the MMIO entity (e.g., a register block).
+  ///
+  /// This initializer is used in standard builds where interposing is not enabled.
+  ///
+  /// - Parameter unsafeAddress: The base memory address for this entity.
   init(unsafeAddress: UInt)
   #endif
 }
