+++ Sources/MMIO/RegisterStorage.swift
@@ -11,26 +11,62 @@
 
 import MMIOVolatile
 
-/// A type that represents the storage of a register value type.
+/// A protocol identifying types suitable for underlying register storage and
+/// capable of performing volatile memory operations.
 ///
-/// The set of types which conform to this protocol is restricted to the set of
-/// volatile operations available on the platform. As such, user code must
-/// _never_ conform new types to this protocol.
+/// `_RegisterStorage` is an internal protocol within Swift MMIO. It restricts
+/// the set of types that can serve as the raw storage for a memory-mapped
+/// register to those for which volatile load and store operations are
+/// explicitly provided. This ensures that all hardware interactions maintain
+/// correct volatile memory semantics, preventing unintended compiler
+/// optimizations.
+///
+/// The conforming types are `UInt8`, `UInt16`, `UInt32`, and `UInt64`. These
+/// correspond to the fixed-width integer types for which the `MMIOVolatile`
+/// C module provides volatile access functions.
+///
+/// This protocol constrains the `Storage` associated type within
+/// ``RegisterValueRaw``, forming a critical part of Swift MMIO's safety
+/// mechanism for hardware interaction.
+///
+/// > Warning: This is an internal protocol. Do not attempt to conform new types
+///   to `_RegisterStorage`.
+///
+/// - SeeAlso: <doc:Volatile-Access>.
 public protocol _RegisterStorage {
-  /// Loads an instance of `self` from the address pointed to by pointer.
+  /// Performs a volatile load of an instance of `Self` from the memory address
+  /// pointed to by `pointer`.
+  ///
+  /// This operation guarantees that the read is not optimized away by the
+  /// compiler and directly accesses the hardware memory location.
+  ///
+  /// - Parameter pointer: An `UnsafePointer` to the memory location (register
+  ///   address) from which to read.
+  /// - Returns: The value read from the memory location.
   static func load(from pointer: UnsafePointer<Self>) -> Self
-  /// Stores an instance of `self` to the address pointed to by pointer.
+
+  /// Performs a volatile store of `value` to the memory address pointed to by
+  /// `pointer`.
+  ///
+  /// This operation guarantees that the write is not optimized away by the
+  /// compiler and directly modifies the hardware memory location.
+  ///
+  /// - Parameter value: The value to write to the memory location.
+  /// - Parameter pointer: An `UnsafePointer` to the memory location (register
+  ///   address) to write to.
   static func store(_ value: Self, to pointer: UnsafeMutablePointer<Self>)
 }
 
 extension UInt8: _RegisterStorage {
-  /// Loads an instance of `self` from the address pointed to by pointer.
+  /// Performs a volatile load of a `UInt8` value from the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func load(from pointer: UnsafePointer<Self>) -> Self {
     mmio_volatile_load_uint8_t(pointer)
   }
 
-  /// Stores an instance of `self` to the address pointed to by pointer.
+  /// Performs a volatile store of a `UInt8` value to the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func store(
     _ value: Self,
@@ -41,13 +77,15 @@ extension UInt8: _RegisterStorage {
 }
 
 extension UInt16: _RegisterStorage {
-  /// Loads an instance of `self` from the address pointed to by pointer.
+  /// Performs a volatile load of a `UInt16` value from the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func load(from pointer: UnsafePointer<Self>) -> Self {
     mmio_volatile_load_uint16_t(pointer)
   }
 
-  /// Stores an instance of `self` to the address pointed to by pointer.
+  /// Performs a volatile store of a `UInt16` value to the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func store(
     _ value: Self,
@@ -58,13 +96,15 @@ extension UInt16: _RegisterStorage {
 }
 
 extension UInt32: _RegisterStorage {
-  /// Loads an instance of `self` from the address pointed to by pointer.
+  /// Performs a volatile load of a `UInt32` value from the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func load(from pointer: UnsafePointer<Self>) -> Self {
     mmio_volatile_load_uint32_t(pointer)
   }
 
-  /// Stores an instance of `self` to the address pointed to by pointer.
+  /// Performs a volatile store of a `UInt32` value to the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func store(
     _ value: Self,
@@ -75,14 +115,17 @@ extension UInt32: _RegisterStorage {
 }
 
 #if arch(x86_64) || arch(arm64)
+// `UInt64` operations are typically available on 64-bit architectures.
 extension UInt64: _RegisterStorage {
-  /// Loads an instance of `self` from the address pointed to by pointer.
+  /// Performs a volatile load of a `UInt64` value from the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func load(from pointer: UnsafePointer<Self>) -> Self {
     mmio_volatile_load_uint64_t(pointer)
   }
 
-  /// Stores an instance of `self` to the address pointed to by pointer.
+  /// Performs a volatile store of a `UInt64` value to the specified memory
+  /// address.
   @inlinable @inline(__always)
   public static func store(
     _ value: Self,
